requirements.txt
flask
google-generativeai
pdfplumber
pandas
openpyxl
